refactor(test): separate fixture package from domain package
