+++ src/test/java/blackjack/domain/card/CardBundleTest.java
@@ -1,8 +1,8 @@
 package blackjack.domain.card;
 
-import static blackjack.domain.fixture.CardRepository.CLOVER4;
-import static blackjack.domain.fixture.CardRepository.CLOVER5;
-import static blackjack.domain.fixture.CardRepository.CLOVER6;
+import static blackjack.fixture.CardRepository.CLOVER4;
+import static blackjack.fixture.CardRepository.CLOVER5;
+import static blackjack.fixture.CardRepository.CLOVER6;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
